Add function utf8decode()
Add function utf8decode() (from the same location as the utf8encode()
but also modified to check for a NULL pointer.

Make the updates of pointers p and utf8 and the int32_t bytes in
decode_json_string() the same order in the surrogate pairs as the single
'\uxxxx' when processing '\u' in strings.

Updated JPARSE_UTF8_VERSION to "1.2.3 2024-10-30".