Improve jparse_bug_report.sh
The script now, in the jparse environment checks, test compiles a simple
file which includes jparse/jparse.h, to see if the dbg, dyn_array and
jparse libraries are installed. This might be useful depending on the
type of problem being reported, and in fact this is what uncovered the
fact that json_utf8.h was not being installed with make install.

The reason it is in the jparse environment checks and not the
compilation checks is that the compilation checks are for compiling
jparse and its tools themselves.