New release - 2.0.0 2024-10-31
The tools jstrencode(1) and jstrdecode(1) have been swapped as
converting a code point to (say) an emoji is actually encoding, not
decoding, according to other sources. Most places that referred to
encode were changed to decode and those that were decode are now encode.
This major change is quite unfortunate but was deemed necessary; as we
have said: sorry (tm Canada :-) ).

The version of the repo, the UTF-8 library, the jparse library,
jstrencode(1), jstrdecode(1) have all been updated to "2.0.0
2024-10-31".

Some duplicate man pages in man/man3 are now symlinks to the source
file, jparse.3. This happened when the initial import to this repo
occurred. Also, the man pages for jstrdecode and jstrencode were swapped
as well as fixed (the terms were incorrect in those man pages wrt the
tool).