Fix comments in json_parse.c
When swapping the encode/decode terminology/tool names the comments for
json_encode() and json_decode() were not swapped. The comments have now
been swapped and also improved.

The Makefile picky rule now uses -8 as in the comments of json_parse.c I
have added emojis to explain how json_decode() does not convert a
Unicode character like 🐉 to its code point and in fact the following is
(for reasons we do not comprehend) perfectly valid JSON:

    { "🔥" "🐉" }

Instead it will do something like converting a '\' to '\\', for example.