Improve README.md
Added table of contents (has to be verified to work on GitHub).

Reordered the sections a bit. The library is moved to right below the
jparse tool whereas the other tools are below that. With the table of
contents this shouldn't be a problem to navigate.

Added some examples (or at least one example) to jstrencode (also gave
more details).