Removed unused utf8decode()
It appears that we will not need the function utf8decode() from
json_utf8.c after all.

Added extra sanity checks in surrogates_to_unicode() that warn if hi or lo
are < 0. This should theoretically never happen but the point is that if it
does happen it is warned about before returning UNICODE_NOT_SURROGATE_PAIR
(these would be triggered in the other check too but we do not warn about those
because the function tests whether it is a surrogate pair or not, so the only
function difference is that if either is < 0 it is warned about as it should
never happen).

Updated JPARSE_UTF8_VERSION to "2.0.3 2024-11-07".