Adjust how data is printed in fprint_line_buf()
After some discussion, part of the change to fprint_line_buf() from
yesterday was rolled back. In particular, the fprint_line_buf() function
should print, as part of debug output, the raw data and not
encoded/decoded strings. Thus if '\a' is encountered it should print
just "\a", not "\\a", and if Ã is encountered then it should print
"\xc3\x83", not "\\xc3\\x83", and so on.

The return value error fix was kept in place and as an extra sanity
check before checking isprint() check isascii(c), though this might not
matter in some or maybe most cases.