Remove '-' from sub-minimal
Although some spell checkers do not recognise it as correct, and
although it might look a bit nicer with the '-' in some cases, it is
still a word without the '-' and some dictionaries will recognise
subminimal but not sub-minimal so I have changed it to be subminimal.

We might call this subminimal because it subtracted subtraction! :-)