Remove -s in jstrencode/jstrdecode, other fixes
Although technically valid, some misleading examples were removed from
the man pages of jstrencode and jstrdecode. Overlooking this simple
fact, and forgetting, the -s option was added in error as it (as the
name suggests) acts on strings only, even if it can look like it is
acting on other data types.

The jstrdecode tool has a new option -d which says to not require
enclosing double quotes in the string. This uses the updated library
code from earlier today, changing -s to -d, with slight modifications.

The jstrencode tool needs to double quote output as that is how JSON
strings must be so it now does that.

Added a simpler version of parse_json() to the parser called
parse_json_str(). It passes in "-" as the filename to parse_json(). This
is unused currently but there are times where it would be useful, though
the persistent problem of error messages expecting filenames would mean
that it would show '-' for stdin.