Update util test code
The util_test.c (which is a copy of util.c but with a defined macro) now
tests the new function open_json_dir_file() (renamed from
open_dir_json_file()), using it on jparse.json. If the file is missing
or is not valid JSON or cannot be opened for some reason make test will
fail.