Validate JSON before decoding JSON in jstrencode
The jstrdecode now makes sure that the string or strings is/are valid
JSON prior to decoding the JSON. It is possible to disable this by use
of the new -j option. Use of the new -J level option sets the JSON
verbosity level.

The jstr_test.sh script uses -j but it would be better (and it might be
looked into in the future) if it did not use it. There is currently no
way to use -J in the jstr_test.sh script but this might be added in the
future.

The jstrdecode version has been updated to "2.1.3 2024-11-18".