Bug fix JSON debug output for numbers
Due to a flow control problem, the floating notation flag was always set
to true if no errors were found and a '.' was found in the string. This
meant that the F flag was always present even for non-floating point
numbers.

The jparse library version and the jparse version have both been
updated.