Add new JSON and err file to bad_loc/
Add new JSON file and error location files
test_jparse/test_JSON/bad_loc/party.json and
test_jparse/test_JSON/bad_loc/party.json.err. This is necessary as it allows
another repo, the mkiocccentry
repo (https://github.com/ioccc-src/mkiocccentry), to not have any problems and
still be able to test the party.json error file. This problem has to do with the
error message has to be exact and rebuilding the error location file was
problematic with the improvement to yyerror() in commit
222c2ec75b2cecc817106ccc025e4aa78a5bd47b.