Fixes and improvements in prep.sh
Add test_jparse/not_a_comment.sh and update test_jparse/prep.sh to use
it.  This will test if any Makefile.local files exist and if any do
it'll warn at the end so that the user of make prep or make release can
be aware of it and be sure it does not skew the results.

The script not_a_comment.sh is from the mkiocccentry repo and was
written by Landon Curt Noll (thanks!).

Fix shellcheck for test_jparse/prep.sh and add script to missing
SH_FILES in test_jparse/Makefile (the reason it was not caught by make
shellcheck is because the script was missing from the Makefile).