Various bug fixes and enhancements
Fix function count_comps(). To explain the fixes, we assume that the path is
"foo///" and the delimiter character is '/': we now remove all but the last
trailing '/'. With this there is one component. In other words, if there are 0
delimiting characters it is 0. If however the string is "foo//foo" the middle
"//" is changed to "/" and it ends up being one component. But if it was
instead "foo//foo/" it would be two because of the trailing '/'. This allows
for properly counting directories at the same time as accounting for files in
the directories. For instance if there is a filename bar in the directory
foo then it is incorrect to say that there are two directories. This could be
deemed problematic in some cases and it might be better to have a boolean to
decide but for now that is not the case.

Updated util test code quite a bit: when no error is encountered in the
functions added in the previous releases (dir_name(), count_dirs() and
sane_relative_path()) we print out the result as well. Also more test cases
were added to test the above change with how components are counted.

Fixed format warnings in various files so that we no longer need -Wno-format.
Fixing this allows for errors and suspect code (both types of issues were
corrected, both in json_parse.c and json_utf8.c).

Updated PARSE_LIBRARY_VERSION to "2.2.6 2025-01-17".
Updated JPARSE_UTF8_VERSION to "2.0.6 2025-01-17".