Improve version details in verge documentation
The man page tries to be a bit more specific (beyond the previous
commit). The code goes back to the form of the other tools.

Added documentation of verge to jparse_utils_README.md. As other matters
are far more pressing the rest of the tools (those that will be
documented in time, hopefully), will be delayed for some time, probably
months at the very least.