New util function mkdirs()
New util function mkdirs() (using mkdir(2)) which acts as mkdir -p with
specific modes (uses chmod(2) as it's not affected by the umask and also
because anything but permissions set with mkdir(2) is undefined). If the first
arg (an int) is -1 (actually < 0) it uses the current working directory to
start out with; but one can pass a file descriptor of the directory to start out
with. The mode is only set on directories that are created (i.e. no error)
because otherwise an already existing directory could have its mode changed.
Just like with mkdir(2) one must be careful with the mode. Of course if one
sets a mode like 0 then trying to work under it would be a problem but that's on
the user. If there is an error in creating a directory then it only aborts if
errno is not EEXIST (already exists) so that it can continue (just like
 mkdir -p).

make clobber in test_jparse/Makefile now removes the test directories
created by util_test (which creates a directory tree of test_jparse/a/b/c
and test_jparse/a).

Updated JPARSE_UTILS_VERSION to "1.0.6 2025-02-05".
Updated UTIL_TEST_VERSION to "1.0.9 2025-02-05.