New util function for FTS library calls
New util function read_fts() which allows one to use fts_open() and
fts_read() in a 're-entrant' way (not strictly true: one always has to pass to
fts_read() the FTS *) an more importantly allowing one to not have to rely
on using fts_open(), checking for NULL, then in a loop using fts_read() etc.
The args are as follows:

    /*
     *  dir     -   char * which is the path to chdir(2) to before opening path but
     *              only if != NULL && *fts == NULL
     *  dirfd   -   if dir == NULL and dirfd > 0, fchdir(2) to it, else don't change
     *              at all
     *  cwd     -   if != NULL set *cwd PRIOR to chdir(dir)
     *  options -   options to pass to fts_open()
     *  fts     -   pointer to pointer to FTS to set to return value of fts_open()
     *  compar  -   if != NULL use it for the compar() function in fts_open(), else
     *              use fts_cmp() (see also fts_rcmp())
     */

One can use the function like:

    FTS *fts = NULL;                    /* FTS stream for fts_open() */
    FTSENT *ent = NULL;                 /* FTSENT for each item from read_fts() */

    ent = read_fts(NULL, -1, NULL, FTS_NOCHDIR | FTS_PHYSICAL, &fts, fts_cmp);
    if (ent == NULL) {
        /* handle error */
    }  else {
        do {
            /* do stuff per entry */
        } while ((ent = read_fts(NULL, -1, NULL, FTS_NOCHDIR | FTS_PHYSICAL, &fts, fts_cmp) != NULL);
    }

or so, san typos.

The function fts_cmp() is like strcmp() with the full path of a file (from
the directory . - see below) and the function fts_rcmp() is the opposite of
strcmp() on the full path.

The function (read_fts()) does check for specific error conditions. For more
details read the comments or the source (until eventually the functions in
util.c are documented better).

Updated JPARSE_REPO_VERSION to "2.2.16 2025-02-13".
Updated UTIL_TEST_VERSION to "1.0.11 2025-02-13".