Add function find_file() that uses read_fts()
This function searches for a file by name based on the parameters. In
particular it can search for basename or full path (relative to the
cwd - see below).

If base is true search by basename.

If depth > 0 it means the file must be at a specific depth.

If count > 0 it means that it has to be the count time the file was seen
(only useful for basename, of course).

If filename is NULL or is an empty string return NULL.

If dir is not NULL try chdir(dir); if it fails or dir is NULL and dirfd
is > -1 then try fchdir(dirfd) (if that fails it is an error but if you
don't want to change the directory you can give NULL and -1).

The compar() callback is passed to fts_open() in read_fts().

The options are options to be passed to read_fts() (which passes them to
fts_read()). If <= 0 it's set to FTS_NOCHDIR and if > 0 FTS_NOCHDIR is
set (this is important so the path will be valid).

The function is:

    extern char const *find_file(char const *filename, char const *dir, int dirfd, int *cwd, bool base,
            int (*compar)(const FTSENT **, const FTSENT **), int options, int count, short int depth);

where the filename arg is the name to look for,
dir is NULL or chdir(dir) fails).

util_test tests this function.

New feature of read_fts() allows to restore the original directory:

    read_fts(NULL, -1, &cwd, -1, NULL, NULL);

If everything is NULL / < 0 except for cwd and *cwd >= 0 it'll try doing
fchdir(*cwd) to restore the original directory (which depending on the
original call to the function can change the directory). After restoring
the directory the FD is closed.